requirements.txt
flask
python-dotenv
flask-cors
requests
reportlab
